This branch was opened against main before parallel audit fixes
landed. Conflicts are all 'PR proposed fix X / main got version Y
of the same fix' — resolved by taking main's version.

+++ crates/evaporchain-singh-heir/src/token.rs
@@ -477,10 +477,10 @@ mod tests {
                 let _ = n.inherit();
                 current_holder = next;
             }
-            let expected = if hops as u32 > init_pow {
+            let expected = if hops > init_pow {
                 0
             } else {
-                1u64 << (init_pow - hops as u32)
+                1u64 << (init_pow - hops)
             };
             proptest::prop_assert_eq!(n.energy, expected);
         }
